Move SSHPTYResizeCoordinator into its own file
Aziz file-organization policy: a major type should live in its own
TypeName.swift rather than appended to CMUXCLI+SSHCommandSupport.swift (which is
an extension CMUXCLI of SSH command-string helpers). Relocate the coordinator
and its two error types to CLI/SSHPTYResizeCoordinator.swift, wired into the
cmux-cli target. No behavior change.
